Move defaultChoice helpers to helpers.
