+++ timmons-group/src/shared-react-components/helpers/helpers.js
@@ -633,3 +633,34 @@ export function convertToLinkFormat(linkFormat, dataNode) {
 
   return link;
 }
+
+/**
+ * The default choice formatter
+ * @param {object} item
+ * @param {object} [options] - options for the choice mapper
+ * @param {string} [options.mappedId] - property to use when mapping the id
+ * @param {string} [options.mappedLabel] - property to use when mapping the label
+ * @returns {Array<object>}
+ */
+export function defaultChoiceFormatter(item, options) {
+  const opt = item || {};
+  const { mappedId, mappedLabel } = options || {};
+  const id = mappedId && opt[mappedId] ? opt[mappedId] : opt.id || opt.streamID;
+  const label = mappedLabel && opt[mappedLabel] ? opt[mappedLabel] : opt.name || opt.label;
+  return { ...item, id, label };
+}
+
+/**
+ * The default parser for fetching choices. Assumes data is an array of objects and a property on response
+ * @param {object} res - response from the fetch
+ * @param {object} [options] - options for the choice mapper
+ * @param {string} [options.mappedId] - property to use when mapping the id
+ * @param {string} [options.mappedLabel] - property to use when mapping the label
+ * @returns {Array<object>}
+ */
+export function defaultChoiceMapper(res, options) {
+  const { data } = res || {};
+  return data?.map((opt) => {
+    return defaultChoiceFormatter(opt, options);
+  });
+}
